% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Data Set for Fish}
\format{
A data frame with 144 rows and 6 columns
\describe{
  \item{RIVER}{Location of fish}
  \item{MILE}{Location of the fish within the river}
  \item{SPECIES}{Species of fish}
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}
  \item{DDT}{DDT level}
  ...
}
}
\source{
{Textbook}
}
\usage{
ddt
}
\description{
144 lines ofmeasurements of fish
}
\keyword{datasets}
