% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{DDT Graph}
\usage{
myddt(df, x = "LENGTH", y = "WEIGHT", SPECIES, col = "RIVER")
}
\arguments{
\item{df}{Dataframe}

\item{x}{x label}

\item{y}{y label}

\item{SPECIES}{Species of fish}

\item{col}{Color}
}
\value{
Detailed plot of data for a specific species
}
\description{
DDT Graph
}
