% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.R, R/myncrve.R
\name{myncrve}
\alias{myncrve}
\title{myncrve}
\usage{
myncrve(mu, sigma, x)

myncrve(mu, sigma, x)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Variance}

\item{x}{X value}
}
\value{
A curve shaded below x

A curve shaded below x
}
\description{
myncrve

myncrve
}
\examples{
\dontrun{myncrve(1,1,1)}
\dontrun{myncrve(1,1,1)}
}
